2025-03-10 18:07 by shun - 채무 Long 타입으로 바꾸기

+++ src/main/java/com/beauty1nside/accnut/dto/DebtDTO.java
@@ -11,7 +11,7 @@ public class DebtDTO {
 	String section;
 	Date registerDate;
 	String creditor;
-	int amount;
+	Long amount;
 	double interest;
 	Date timeLimit;
 	Date prearrangementDueDate;
